hunter-notes-ai-service/requirements.txt
Flask
Pillow
# Future: Agno specific libraries, e.g., tensorflow, pytorch, opencv-python, etc.